python/requirements.txt
# Core
python-dotenv>=1.0.0
structlog>=24.1.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async
asyncio>=3.4.3
aiofiles>=23.2.1

# HTTP
httpx>=0.26.0
tenacity>=8.2.3

# AWS
boto3>=1.34.0
aioboto3>=12.3.0

# Anthropic
anthropic>=0.40.0

# Database
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# PDF Conversion
playwright>=1.41.0

# CLI (for local testing)
click>=8.1.7
rich>=13.7.0

# Web framework (for health check endpoint)
fastapi>=0.109.0
uvicorn>=0.27.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-mock>=3.12.0
respx>=0.20.2
